csv: re-define QuoteStyle and Terminator types.
Re-exporting the QuoteStyle and Terminator types through csv from csv-core
causes csv-core to be a public dependency of csv, which is bad juju.
Instead, redefine the types in csv.
